Create a wrapper component for the home page to prevent any fetch requests when the user isn't logged in
