+++ .netlify/edge-functions/manifest.json
@@ -1,1 +1,1 @@
-{"functions":[{"function":"next-dev","name":"netlify dev handler","path":"/*","generator":"@netlify/next-runtime@4.39.3"}],"version":1}
+{"functions":[{"function":"next-dev","name":"netlify dev handler","path":"/*","generator":"@netlify/next-runtime@4.39.4"}],"version":1}
