+++ components/layout/Navbar/NavBar.module.css
@@ -1,7 +1,7 @@
 .main {
   display: flex;
   width: 100%;
-  height: 100px;
+  height: var(--navbar-height);
   justify-content: center;
   align-items: center;
   color: #141349;
@@ -12,7 +12,7 @@
   display: grid;
   grid-template-columns: 0.5fr 1fr;
   list-style-type: none;
-  height: 100px;
+  height: var(--navbar-height);
   align-items: center;
   padding-left: 5vw;
   background-color: white;
