+++ styles/globals.css
@@ -6,6 +6,7 @@
   --blue: #141349;
   --light-blue: #E0DFFF;
   --gray: #D9D9D9;
+  --navbar-height: 100px;
 }
 
 * {
